...G...... [ZBXNEXT-801] coding style fixes and formatting
Implemented by sasha.


git-svn-id: svn://svn.zabbix.com/branches/dev@28106 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-801-2.0/src/libs/zbxcomms/telnet.c
@@ -65,7 +65,7 @@ static ssize_t	telnet_socket_read(ZBX_SOCKET socket_fd, void *buf, size_t count)
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __function_name);
 
-	while (ZBX_TCP_ERROR == (rc = ZBX_TCP_READ(socket_fd, buf, (int)count)))
+	while (ZBX_TCP_ERROR == (rc = ZBX_TCP_READ(socket_fd, buf, count)))
 	{
 		error = zbx_sock_last_error();	/* zabbix_log() resets the error code */
 		zabbix_log(LOG_LEVEL_DEBUG, "%s() rc:%d errno:%d error:[%s]",
@@ -102,7 +102,7 @@ static ssize_t	telnet_socket_write(ZBX_SOCKET socket_fd, const void *buf, size_t
 
 	zabbix_log(LOG_LEVEL_DEBUG, "In %s()", __function_name);
 
-	while (ZBX_TCP_ERROR == (rc = ZBX_TCP_WRITE(socket_fd, buf, (int)count)))
+	while (ZBX_TCP_ERROR == (rc = ZBX_TCP_WRITE(socket_fd, buf, count)))
 	{
 		error = zbx_sock_last_error();	/* zabbix_log() resets the error code */
 		zabbix_log(LOG_LEVEL_DEBUG, "%s() rc:%d errno:%d error:[%s]",
